reorganize & refactor tests
